CSV plot in the seperate program
